pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "language-tool-python>=2.9.4",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.1",
    "spacy>=3.8.7",
    "weasyprint>=66.0",
    "werkzeug>=3.1.3",
]
